Case
Components imported to App.js did not specify their complete filepath, just the component folder without "Index.js" at the end of the path. React apparently automatically looks for the lower case "index.js" not uppercase "Index.js," and resaving those files with the proper case is not reflected during the build process for some reason, so I simply went through the imports in App.js and specified the filepaths with the correct cases.